Enhance remote connection behavior. Add a buffered channel as cache buffer in front of output. In case the client does not receive the message gets buffered so a state is always there if the client receives again. If the buffer can't be written it must be full so the message is received to clean the buffer and the current message is then sent. This keeps the latest state up to date.
For led I put a buffer size of 4 since I know there are always 4 status updates for each LED update.
But I am not sure if this is really deterministic.
It would be easier to intercept the led.State channel which contains a full state of all leds instead of handling led changes per led.